Flag-gate the complete topology
Put the "complete topology" behavior behind a CLI flag, "-k" or
"--complete". The "k" is a reference to the common graph theory notation
for a complete graph.

Used this flag to measure the impace of enabling the complete topology.

Procedure:
1. Render killeroo with static assignment and 100 workers.
2. Measure net bytes sent and loss rate at 10s.

Results:

trial  |  -k  |  loss  |  bytes sent |
-------+------+--------+-------------+
1      |  no  | 0.0473 | 745M        |
2      |  no  | 0.0441 | 774M        |
3      |  no  | 0.0419 | 740M        |
4      | yes  | 0.0203 | 698M        |
5      | yes  | 0.0120 | 687M        |
6      | yes  | 0.0304 | 652M        |
7      | yes  | 0.0165 | 691M        |

So we lose a little on bytes sent, and win on loss.